fix: check if line exists to append

+++ src/controller.ts
@@ -1,7 +1,11 @@
 import { appendFileSync, writeFileSync } from "fs";
+import { checkLineExistsInFile } from "./helpers";
 
 export function generateController(component: string) {
-  writeFileSync(`src/controllers/${component.toLocaleLowerCase()}.ts`, 'import { api } from "../api";\n');
+  if (checkLineExistsInFile("src/controllers/index.ts", component)) {
+    return;
+  }
 
+  writeFileSync(`src/controllers/${component.toLocaleLowerCase()}.ts`, 'import { api } from "../api";\n');
   appendFileSync("src/controllers/index.ts", `import "./${component.toLocaleLowerCase()}";\n`);
 }
